+++ src/components/Feed/Feed.styles.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+export const VideoSlider = styled('section')`
+  position: relative;
+  inset: 0;
+  width: 260px;
+  height: 562px;
+  border-radius: 26px;
+  display: flex;
+  flex-direction: column;
+  overflow-y: scroll;
+  scroll-snap-type: y mandatory;
+`