+++ src/components/Phone/Phone.tsx
@@ -1,11 +1,12 @@
+
+import { ReactElement } from 'react'
+
 import { PhoneContainStyled, PhoneStyled } from './Phone.styles'
 
-export function Phone() {
+export function Phone({ children }: Readonly<{ children: ReactElement }>) {
   return (
-    <>
-      <PhoneStyled />
-      <PhoneContainStyled></PhoneContainStyled>
-    </>
+    <PhoneStyled>
+      <PhoneContainStyled>{children}</PhoneContainStyled>
+    </PhoneStyled>
   )
 }
-