+++ src/components/Video/video.styles.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+export const VideoStyled = styled('video')`
+  max-width: 260px;
+  min-height: 100%;
+  aspect-ratio: 9/16;
+  object-fit: cover;
+  scroll-snap-align: center;
+`
+