+++ src/components/index.ts
@@ -1,2 +1,5 @@
+
 export * from './Phone'
+export * from './Feed'
+export * from './Video'