+++ src/theme/GlobalStyle.tsx
@@ -1,3 +1,4 @@
+
 import { createGlobalStyle } from 'styled-components'
 
 export const GlobalStyle = createGlobalStyle`
@@ -20,7 +21,11 @@ export const GlobalStyle = createGlobalStyle`
         height: 600px;
         display: flex;
         align-items: center;
-        justify-content: center;
+        justify-content: center;      
     }
+
+    ::-webkit-scrollbar {
+    display: none;
+  }
 `