+++ changelog.md
@@ -1,4 +1,13 @@
+
 # CHANGELOG
+
+## 9-3-2024 v.0.0.2
+
+## Features.
+
+- add videos on feed
+- change scaffolding
+- managment play and pause videos
 
 ## 7-3-2024 v.0.0.1
 
@@ -13,4 +22,3 @@
 ## Features.
 
 - Add initial config
-